Separate compile-time template processing
Amp-Thread-ID: https://ampcode.com/threads/T-01a060ab-1198-76c8-8e43-674c2971257f
